+++ foodkept/Pages/User/Login.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FoodKept.Pages.LoginModel
+@model FoodKept.Pages.User.LoginModel
 @{
 }
