+++ foodkept/Pages/User/Profile.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FoodKept.Pages.ProfileModel
+@model FoodKept.Pages.User.ProfileModel
 @{
     ViewData["Title"] = "Profile";
 }
