+++ foodkept/Pages/User/Register.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FoodKept.Pages.RegisterModel
+@model FoodKept.Pages.User.RegisterModel
 @{
     ViewData["Title"] = "Register";
 }
